pyproject.toml
[project]
name = "roondal-api"
version = "2.0.0"
description = "Roondal Mattermost Integration API"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115",
    "uvicorn[standard]>=0.30",
    "sqlalchemy[asyncio]>=2.0",
    "asyncmy>=0.2",
    "asyncpg>=0.30",
    "pydantic-settings>=2.0",
    "httpx>=0.27",
    "google-api-python-client>=2.100",
    "google-auth>=2.20",
    "google-auth-oauthlib>=1.0",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0",
]
